+++ lib/mempool/src/lib.rs
@@ -1,19 +1,24 @@
 mod reth;
-mod stream;
-mod traits;
+pub use reth::RethPool;
 
-pub use crate::reth::RethPool;
-use reth_chainspec::{ChainSpecProvider, EthereumHardforks};
-use reth_storage_api::StateProviderFactory;
-pub use reth_transaction_pool::error::PoolError;
-pub use reth_transaction_pool::{
-    CanonicalStateUpdate, EthPooledTransaction, NewSubpoolTransactionStream, NewTransactionEvent,
-    PoolUpdateKind, TransactionPool as RethTransactionPool,
-    TransactionPoolExt as RethTransactionPoolExt,
-};
+mod stream;
 pub use stream::{BestTransactionsStream, ReplayTxStream, TxStream, best_transactions};
+
+mod traits;
 pub use traits::L2TransactionPool;
 
+mod transaction;
+pub use transaction::L2PooledTransaction;
+
+// Re-export some of the reth mempool's types.
+pub use reth_transaction_pool::error::PoolError;
+pub use reth_transaction_pool::{
+    CanonicalStateUpdate, NewSubpoolTransactionStream, NewTransactionEvent, PoolUpdateKind,
+    TransactionPool as RethTransactionPool, TransactionPoolExt as RethTransactionPoolExt,
+};
+
+use reth_chainspec::{ChainSpecProvider, EthereumHardforks};
+use reth_storage_api::StateProviderFactory;
 use reth_transaction_pool::blobstore::NoopBlobStore;
 use reth_transaction_pool::validate::EthTransactionValidatorBuilder;
 use reth_transaction_pool::{CoinbaseTipOrdering, PoolConfig};