+++ lib/rpc_api/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod debug;
 pub mod eth;
 pub mod filter;
+pub mod net;
 pub mod ots;
 pub mod pubsub;
 pub mod types;